requirements.txt
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
openai>=1.30.0
PyGithub>=2.3.0
plotly>=5.22.0
kaleido>=0.2.1
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
requests>=2.32.0
pandas>=2.2.0
networkx>=3.3
pyvis>=0.3.2
Jinja2>=3.1.4
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
